+++ src/App.js
@@ -6,7 +6,6 @@ import './App.css';
 import Home from './components/Home'
 import QuestionForm from './components/QuestionForm'
 import QuestionList from './components/QuestionList'
-import CommentForm from './components/CommentForm'
 
 function App() {
   return (
@@ -30,7 +29,6 @@ function App() {
           <Route exact path='/' component={Home} />
           <Route path='/questionlist' component={QuestionList} />
           <Route path='/questionform' component={QuestionForm} />
-          <Route path='/commentform' component={CommentForm} />
         </Switch>
       </div>
     </Router>
